+++ src/action/interact/confirm/pay.ts
@@ -12,7 +12,12 @@ export interface IPayAssetTransaction {
     transactionNumber: string;
     // 確定時の決済方法区分指定(2023-08-29~)
     object?: {
-        paymentMethod?: { identifier?: string };
+        paymentMethod?: {
+            /**
+             * 決済方法区分
+             */
+            identifier?: string;
+        };
     };
 }
 export type IObject = IPayAssetTransaction[];
@@ -20,7 +25,6 @@ export type IResult = any;
 export interface IAttributes extends ConfirmActionFactory.IAttributes<IObject, IResult> {
     // 資産取引化(2024-03-11~)
     instrument: IInstrument;
-    // potentialActions?: any;
     purpose: IPurpose;
 }
 /**