+++ src/action/trade/order.ts
@@ -2,7 +2,7 @@ import * as ActionFactory from '../../action';
 import { ActionType } from '../../actionType';
 import { ISimpleOrder } from '../../order';
 import { TransactionType } from '../../transactionType';
-import { IAttributes as IGivePointAwardActionAttributes } from '../transfer/give/pointAward';
+// import { IAttributes as IGivePointAwardActionAttributes } from '../transfer/give/pointAward';
 import { IAttributes as ISendOrderActionAttributes } from '../transfer/send/order';
 
 export type IAgent = ActionFactory.IParticipant;
@@ -11,11 +11,7 @@ export type IObject = ISimpleOrder;
 export type IResult = any;
 export type ISendOrderPotentialAction = Pick<ISendOrderActionAttributes, 'potentialActions'>;
 export interface IPotentialActions {
-    /**
-     * ポイント付与アクション
-     * 現時点で複数口座にポイントを付与することはないが、可能性もこめてリストで持っておく
-     */
-    givePointAward?: IGivePointAwardActionAttributes[];
+    // givePointAward?: IGivePointAwardActionAttributes[]; // 廃止(2024-03-12~)
     // pay?: IPayActionAttributes[]; // 廃止(2023-02-03~)
     // sendOrder?: ISendOrderActionAttributes;
     sendOrder?: ISendOrderPotentialAction; // optimize(2024-01-16~)