+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 資産取引確定アクションのpotentialActionsを廃止
 - 決済取引確定アクションのinstrumentを廃止
 - アクションのIParticipantからidentifierを明示的に除外
+- 予約取引確定アクションのinstrumentを廃止
 
 ### Deprecated